+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2024-06-05
+
+- yellowstone-grpc-geyser-1.14.4+solana.1.18.15
+
+### Features
+
+- geyser: add metric `subscriptions_total` ([#355](https://github.com/rpcpool/yellowstone-grpc/pull/355))
+
 ## 2024-06-02
 
 - yellowstone-grpc-client-1.15.0+solana.1.18.15